src/requirements.txt
albumentations==1.4.17
evaluate==0.4.3
ipykernel==6.29.5
jupyter==1.1.1
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
protobuf==3.20.3
PyGithub==1.56
tensorflow==2.7.0
transformers==4.21.0
